Split rest client script
